Format values from Satoshi to BTC.

+++ src/test/java/com/hamzaikine/bitcoindashboard/AddressTest.java
@@ -18,25 +18,24 @@ import static org.junit.Assert.*;
  * @author hamzaikine
  */
 public class AddressTest {
-    
-    
-    Address ad = new Address("","1F1tAaz5x1HUXrCNLbtMDqcw6o5GNn4xqX",1,2,3,4,null);
-    
+
+    Address ad = new Address("", "1F1tAaz5x1HUXrCNLbtMDqcw6o5GNn4xqX", 1, 2, 3, 4, null);
+
     public AddressTest() {
     }
-    
+
     @BeforeClass
     public static void setUpClass() {
     }
-    
+
     @AfterClass
     public static void tearDownClass() {
     }
-    
+
     @Before
     public void setUp() {
     }
-    
+
     @After
     public void tearDown() {
     }
@@ -47,34 +46,34 @@ public class AddressTest {
         String expResult = "1F1tAaz5x1HUXrCNLbtMDqcw6o5GNn4xqX";
         String result = ad.getAddress();
         assertEquals(expResult, result);
-        
+
     }
 
     @Test
     public void testGetTotalReceived() {
         System.out.println("getTotalReceived");
-        long expResult = 1;
-        long result = ad.getTotalReceived();
+        double expResult = 1;
+        double result = ad.getTotalReceived();
         assertEquals(expResult, result);
-       
+
     }
 
     @Test
     public void testGetTotalSent() {
         System.out.println("getTotalSent");
-        long expResult = 2;
-        long result = ad.getTotalSent();
+        double expResult = 2;
+        double result = ad.getTotalSent();
         assertEquals(expResult, result);
-        
+
     }
 
     @Test
     public void testGetFinalBalance() {
         System.out.println("getFinalBalance");
-        long expResult = 3;
-        long result = ad.getFinalBalance();
+        double expResult = 3;
+        double result = ad.getFinalBalance();
         assertEquals(expResult, result);
-        
+
     }
 
     @Test
@@ -83,7 +82,7 @@ public class AddressTest {
         int expResult = 4;
         int result = ad.getTxCount();
         assertEquals(expResult, result);
-       
+
     }
 
     @Test
@@ -92,7 +91,7 @@ public class AddressTest {
         List<Transaction> expResult = null;
         List<Transaction> result = ad.getTransactions();
         assertEquals(expResult, result);
-       
+
     }
-    
+
 }
